requirements.txt
Django~=4.2
python-dotenv~=1.0.1
djangorestframework~=3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
aiogram~=3.15.0
djangorestframework-simplejwt==5.3.1
celery>=5.2.7
redis>=4.5.1
drf-yasg==1.21.8
django-cors-headers
coverage==7.6.8
requests==2.32.3
pytz==2024.2
django-celery-beat==2.7.0
flake8==6.0.0
